refactor(suite): fix TroubleshootingTipsItem type-only circular dep
