+++ packages/suite/src/components/suite/PrerequisitesGuide/DeviceBootloader.tsx
@@ -4,7 +4,7 @@ import { TroubleshootingTips } from 'src/components/suite/troubleshooting/Troubl
 import { useDevice } from 'src/hooks/suite/useDevice';
 import { getHowToGetFromBootloaderInstructionsMap } from 'src/utils/device/bootloader';
 
-import { type TroubleshootingTipsItem } from '../troubleshooting/TroubleshootingTips';
+import { type TroubleshootingTipsItem } from '../troubleshooting/TroubleshootingTipsItem';
 import { UpdateGoToSettingsDescription } from '../troubleshooting/tips/UpdateGoToSettingsDescription';
 
 /* User connected the device in bootloader mode, but in order to continue it needs to be in normal mode */
